requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.7.0
idna==3.10
python-dateutil==2.9.0.post0
python-whois==0.9.5
requests==2.32.5
six==1.17.0
urllib3==2.5.0
